pyproject.toml
[tool.uv]
package = true

[project]
name = "hackathon-todo-cli"
version = "0.1.0"
dependencies = [
  "sqlmodel",
  "python-dotenv"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]